Permissions: Added an activity for permission request.
If the user does not provide permissions to the application
for audio camera and connection monitoring the webrtc code
will crash. With this pre activity we have a way to avoid that.

Not on expert on android but its is quite weird that the permission
request is not synchronous....oh well.