Replaced websocket library with standard java one. Part 2.
The thread context of the websocket message is different than
the main one leading to funny races. Do not know if the race
already happened with the other websocket library but i would
guess yes. I think the reason this was happening is due to the
very convoluted code around interfaces and delegates from the main
activity and web socket janus communications