PeerConnectionClient: No longer a singleton.
Previously we had the class as a singleton with a minimal constructor.
Now we have state per instance and it should be easier to grasp. I think
the author of this code must have been quite a genious, in the good sense,
because i cannot understand how he was able to reason all the moving parts
into something working.