peerConnectionClient: Removed videoSource and audioSource member variables.
The member variables can be replaced by local variables in the places
where they are needed and set. Helps simplify the state of the class

Also no need to dispose them explicitly as the peerConnection will
take care of disposing of tracks and sources and media streams. All
except the capturer